Moved state management outside of LeoProvider.
